% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lmszscore.R
\name{lmszscore}
\alias{lmszscore}
\title{Calculate z scores}
\usage{
lmszscore(mydf, sexcat = 0, sex = "sex", age = "age", var = "bmi",
  lmsref)
}
\arguments{
\item{mydf}{data frame}

\item{sexcat}{coding of sex for male}

\item{sex}{column name of sex column e.g. "sex"}

\item{age}{column name of age column e.g. "age"}

\item{var}{column name of item e.g. "bmi"}

\item{lmsref}{LMS reference data. Excepts sex unqiue data and needs columns named "age", "L", "M", "S"}
}
\value{
The function returns a vector containing z scores
}
\description{
Calculate z scores using LMS values
}
\details{
´lmszscore´ calculates growth and BMI z scores using LMS reference values
}
\examples{
mydf <- data.frame(  age = c(10.2,11.4,9.7,2.3,14.2,18.2,19.4,13.0),
                     sex = sample(c("male","female"), size = 8, replace = TRUE),
                     bmi = c(20.1,27.3,15.0,18.2,17.4,23.9,21.2,17.9))

lmsref <- data.frame(age = seq(1,20,0.5),
                        L = seq(0.1,-2,length.out = 39),
                        M = seq(13,22,length.out = 39),
                        S = seq(0.09,0.2,length.out = 39))

lmszscore(mydf = mydf, sexcat = "female", sex = "sex", age = "age", var = "bmi" , lmsref = lmsref)

}
\author{
Florian Halbeisen
}
